+++ tests/debugging_tools.cpp
@@ -4,7 +4,7 @@ void stringDifference(std::string string1, std::string string2)
 {
     Catch::cout() << "String 1 len: " << string1.length() << "; string 2 len: " << string2.length() << std::endl;
     
-    int max = std::max(string1.length(), string2.length());
+    size_t max = std::max(string1.length(), string2.length());
     for (int i = 0; i < max; i++)
     {
         if (string1[i] != string2[i])
